requirements.txt
numpy>=1.26.0
pandas>=2.1.0
yfinance>=0.2.31
ta>=0.10.2
matplotlib>=3.8.0
scipy>=1.11.3
pytest>=7.4.0
python-dotenv>=1.0.0
backtrader>=1.9.78.123 